refactor: moved the files to different names that match the common casing for components
